+++ test/test_speed_pool.cpp
@@ -13,12 +13,12 @@ int UNI_DATA_SIZE = -1;
 const int TEST_TIME = 3e4;
 
 int TOT_CNT = 0;
-double MAX_DIFF = -1, TOT_DIFF = 0;
+long double MAX_DIFF = -1, TOT_DIFF = 0;
 
 
-std::pair<double, double> swap_handle(Pool<true> * pool, bool zeroToOne, double amountIn) {
-    static double SQPRL = uint160("4295128740").X96ToDouble();
-    static double SQPRR = uint160("1461446703485210103287273052203988822378723970341").X96ToDouble();
+std::pair<FloatType, FloatType> swap_handle(Pool<true> * pool, bool zeroToOne, long double amountIn) {
+    static FloatType SQPRL = uint160("4295128740").X96ToDouble();
+    static FloatType SQPRR = uint160("1461446703485210103287273052203988822378723970341").X96ToDouble();
     return swap(pool, zeroToOne, amountIn, zeroToOne ? SQPRL : SQPRR, false);
 }
 
@@ -32,13 +32,13 @@ std::pair<int256, int256> raw_swap_handle(Pool<false> * pool, bool zeroToOne, in
 struct testcase{
     bool zeroToOne;
     int256 raw_amount;
-    double amount;
-    std::pair<double, double> result;
-    double dev;
+    FloatType amount;
+    std::pair<FloatType, FloatType> result;
+    FloatType dev;
     int repeat;
 };
 vector<testcase> tc;
-vector<pair<double, double>> result;
+vector<pair<FloatType, FloatType>> result;
 
 void generateFromExpon(Pool<false> * pool) {
     for(uint160 amount = 10; amount < uint160("94033269757636"); (amount *= 14) /= 10) {
@@ -88,10 +88,10 @@ WETH    (0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2) (1): 6530822335762893455121
 USD Coin(0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48) (0): 94033269757636
 */
 
-tuple<std::string, std::string, std::string, std::string> split(double a, double b) {
+tuple<std::string, std::string, std::string, std::string> split(long double a, long double b) {
     static char buffer0[50], buffer1[50];
-    sprintf(buffer0, "%.1lf", a);
-    sprintf(buffer1, "%.1lf", b);
+    sprintf(buffer0, "%.1Lf", a);
+    sprintf(buffer1, "%.1Lf", b);
     std::string a0 = "", a1 = "", b0 = "", b1 = "";
     std::string raw0 = buffer0;
     std::string raw1 = buffer1;
@@ -149,11 +149,11 @@ int main(){
     cerr << "done generated data cnt = " << UNI_DATA_SIZE << endl;
 
 
-    // TestCertainPoint(pool, pool_float, int256("94033269757636"), 1);
+    // TestCertainPoint(pool, pool_float, int256("10"), 1);
     // return 0;
 
     cerr << "run" << endl;
-    double timer[UNI_DATA_SIZE];
+    long double timer[UNI_DATA_SIZE];
     for(int t = 0; t < UNI_DATA_SIZE; t++) {
         timer[t] = clock();
         for(int i = 0; i < TEST_TIME; i++) {
@@ -165,12 +165,12 @@ int main(){
     cerr << "run done" << endl;
 
     for(int i = 0; i < UNI_DATA_SIZE; i++) {
-        pair<double, double> ret1 = result[i], ret0 = tc[i].result;
+        pair<FloatType, FloatType> ret1 = result[i], ret0 = tc[i].result;
         // cout << ret0.first << " " << ret1.first << endl;
         // cout << ret0.second << " " << ret1.second << endl;
         bool zeroToOne = tc[i].zeroToOne;
-        double amountSpecified = tc[i].amount;
-        double diffe;
+        FloatType amountSpecified = tc[i].amount;
+        FloatType diffe;
 
         diffe = max(fabs(  (ret1.second - ret0.second) / std::max(fabs(ret0.second), fabs(ret1.second))  ),
                 fabs(  (ret1.first - ret0.first) / std::max(fabs(ret0.first), fabs(ret1.first))  ));
@@ -186,7 +186,7 @@ int main(){
         // if(diffe < 0.000001 || (ret1.first > -100000 && ret1.first < 100000) || (ret1.second > -100000 && ret1.second < 100000)) ; else {
         //     static char buffer[1000];
         //     sprintf(buffer, "\n\n================================================= FAIL ============================================\n"
-        //                     "%.30lf %.30lf\n%.30lf %.30lf\n",
+        //                     "%.30Lf %.30Lf\n%.30Lf %.30Lf\n",
         //             ret0.first, ret0.second, ret1.first, ret1.second);
         //     std::cerr << buffer << std::endl;
         // }
@@ -199,10 +199,10 @@ int main(){
             std::tie(a0, a1, b0, b1) = split(tc[t].result.first, result[t].first);
         else
             std::tie(a0, a1, b0, b1) = split(tc[t].result.second, result[t].second);
-        printf("%d | % 24.0lf | % 6.0lf | %.8lf | %s%s\033[41;30m%s\033[0m | %s\033[41;30m%s\033[0m \n", tc[t].zeroToOne, tc[t].amount, timer[t], tc[t].dev, whitespace(25 - (a0.size() + a1.size())).c_str(), a0.c_str(), a1.c_str(), b0.c_str(), b1.c_str());
+        printf("%d | % 24.0Lf | % 6.0Lf | %.8Lf | %s%s\033[41;30m%s\033[0m | %s\033[41;30m%s\033[0m \n", tc[t].zeroToOne, tc[t].amount, timer[t], tc[t].dev, whitespace(25 - (a0.size() + a1.size())).c_str(), a0.c_str(), a1.c_str(), b0.c_str(), b1.c_str());
     }
     //printf("======================== Timer ==========================\n");
-    printf("Maximum mistake  = %.30lf\n", MAX_DIFF);
+    printf("Maximum mistake  = %.30Lf\n", MAX_DIFF);
     // printf("Average mistake  = %.30f\n", TOT_DIFF / TOT_CNT);
 
     return 0;
